requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
azure-ai-vision-imageanalysis==1.0.0
python-multipart==0.0.9
azure-storage-blob==12.24.0
openai==1.57.0
azure-ai-documentintelligence==1.0.0
